refactor file directory
