+++ lib/models/RideObject.dart
