+++ lib/pages/Login.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:carriage_rider/AuthProvider.dart';
+import 'package:carriage_rider/providers/AuthProvider.dart';
 
 class Login extends StatefulWidget {
   @override
