+++ lib/pages/Request_Ride_Loc.dart
@@ -1,13 +1,13 @@
-import 'package:carriage_rider/Request_Ride_Time.dart';
+import 'package:carriage_rider/pages/Request_Ride_Time.dart';
 import 'package:flutter/material.dart';
-import 'package:carriage_rider/Home.dart';
-import 'package:carriage_rider/RideObject.dart';
+import 'package:carriage_rider/pages/Home.dart';
+import 'package:carriage_rider/models/RideObject.dart';
 import 'package:flutter_typeahead/flutter_typeahead.dart';
-import 'package:carriage_rider/AuthProvider.dart';
-import 'package:carriage_rider/LocationsProvider.dart';
-import 'package:carriage_rider/app_config.dart';
+import 'package:carriage_rider/providers/AuthProvider.dart';
+import 'package:carriage_rider/providers/LocationsProvider.dart';
+import 'package:carriage_rider/utils/app_config.dart';
 import 'package:provider/provider.dart';
-import 'package:carriage_rider/TextThemes.dart';
+import 'package:carriage_rider/utils/TextThemes.dart';
 
 class RequestRideLoc extends StatefulWidget {
   final RideObject ride;
